+++ src/libs/layer.ts
@@ -70,3 +70,21 @@ export function changeProperty(layer: LayerSpecification, group: keyof LayerSpec
     }
   }
 }
+
+export function layerPrefix(name: string) {
+  return name.replace(' ', '-').replace('_', '-').split('-')[0]
+}
+
+export function findClosestCommonPrefix(layers: LayerSpecification[], idx: number) {
+  const currentLayerPrefix = layerPrefix(layers[idx].id)
+  let closestIdx = idx
+  for (let i = idx; i > 0; i--) {
+    const previousLayerPrefix = layerPrefix(layers[i-1].id)
+    if(previousLayerPrefix === currentLayerPrefix) {
+      closestIdx = i - 1
+    } else {
+      return closestIdx
+    }
+  }
+  return closestIdx
+}
